Testing new beach themed background colors for the headers.
Also switched the h5 tagged items to p with a class of location.
Switched the h4 items to h5 (these are the parents) and switched a
dupicate h3 to h4.  So now h2-h4 are header items that extented the
width of the app and h5 is the parent item.  This should be complient
with html accessibility requirements.